rename ngo to scrub-file
