+++ src/scrub-file.ts
